pom cleanup, added EXAMPLE1_PROJECTOR_HANDLER
