+++ crabzilla-core/src/test/java/io/github/crabzilla/example1/example1.kt
@@ -3,11 +3,10 @@ package io.github.crabzilla.example1
 import java.time.Instant
 import java.util.*
 
-fun subDomainName() = "example1"
-
 enum class CommandHandlers {
   CUSTOMER
 }
+
 
 data class CustomerSummary(val id: String, val name: String, val isActive: Boolean)
