+++ crabzilla-pg-client/src/test/java/io/github/crabzilla/pgclient/example1/example1.kt
@@ -1,0 +1,45 @@
+package io.github.crabzilla.pgclient.example1
+
+import io.github.crabzilla.DomainEvent
+import io.github.crabzilla.example1.CustomerActivated
+import io.github.crabzilla.example1.CustomerCreated
+import io.github.crabzilla.example1.CustomerDeactivated
+import io.github.crabzilla.pgclient.PgClientEventProjectorIT
+import io.github.crabzilla.pgclient.ProjectorHandler
+import io.github.crabzilla.pgclient.runPreparedQuery
+import io.reactiverse.pgclient.PgConnection
+import io.reactiverse.pgclient.Tuple
+import io.vertx.core.AsyncResult
+import io.vertx.core.Future
+import io.vertx.core.Handler
+
+val EXAMPLE1_PROJECTOR_HANDLER: ProjectorHandler =
+  { pgConn: PgConnection, targetId: Int, event: DomainEvent, handler: Handler<AsyncResult<Void>> ->
+
+    PgClientEventProjectorIT.log.info("event {} ", event)
+
+    // TODO can I compose here? https://vertx.io/docs/vertx-core/java/#_sequential_composition
+    val future: Future<Void> = Future.future<Void>()
+    future.setHandler(handler)
+
+    when (event) {
+      is CustomerCreated -> {
+        val query = "INSERT INTO customer_summary (id, name, is_active) VALUES ($1, $2, $3)"
+        val tuple = Tuple.of(targetId, event.name, false)
+        pgConn.runPreparedQuery(query, tuple, future)
+      }
+      is CustomerActivated -> {
+        val query = "UPDATE customer_summary SET is_active = true WHERE id = $1"
+        val tuple = Tuple.of(targetId)
+        pgConn.runPreparedQuery(query, tuple, future)
+      }
+      is CustomerDeactivated -> {
+        val query = "UPDATE customer_summary SET is_active = false WHERE id = $1"
+        val tuple = Tuple.of(targetId)
+        pgConn.runPreparedQuery(query, tuple, future)
+      }
+      else -> PgClientEventProjectorIT.log.info("${event.javaClass.simpleName} does not have any event projector handler")
+    }
+
+    PgClientEventProjectorIT.log.info("finished event {} ", event)
+  }
